fixed release script

+++ release.sh
@@ -1,5 +1,7 @@
 #/bin/bash
 set -eux
-rm -r dist
+if [-d "dist"]; then
+    rm -r dist
+fi
 python3 setup.py sdist bdist_wheel
 twine upload dist/* -u m1n3rva -p "sndoA1v6ao87jKTAU45i"
